requirements.txt
fastapi
python-multipart
fitz-python
pdfplumber
pytesseract
Pillow
pdf2image
python-magic
